feat: moved reply privately extra to ui
